refactor: extract sub-components from OrderFuel
Extracted four sub-components from OrderFuel.tsx to improve maintainability and readability:
- PriceTrendIndicator.tsx
- VehicleSelector.tsx
- EmergencyToggle.tsx
- DeliveryInstructions.tsx
